Separate functions to render three types of cells: empty, nought & cross
Rather than write everything in one component and create a monster of a function
to read, the different cell types were split into their own functions.  The cell
function to call is based on the contents of that location in the app-state.

To test the rendering of each cell, the on-click value in the cell-empty
function was manually changed to :nought when testing the cell-nought function
and :cross when testing the cell-cross function.