Updated simple logging output
Using the on-js-reload function to print out the game board state on ever
change, to give a quick way to check what is happening with the state in the
browser developer console.

Removed the printing of the game board from the cell-empty :on-click section as
it is now covered by the on-js-reload.

Updated cell-empty and computer-move functions to log who has taken which move
and where.